Remove unnecessary PartsField methods from Hidden (#373)
Extract input-tag methods into InputTagMethodsTrait, shared between InputField
and the new InputTagField base. Hidden now extends InputTagField instead of
InputField, removing template/label/hint/error/container methods from its API.
