+++ src/Field/Base/InputTagField.php
@@ -1,0 +1,26 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Form\Field\Base;
+
+use Yiisoft\Form\Field\Base\InputData\InputDataWithCustomNameAndValueTrait;
+use Yiisoft\Form\Field\Base\InputTag\InputTagMethodsTrait;
+
+/**
+ * Base class for fields that render a simple input tag without template/label/hint/error/container features.
+ *
+ * @see PartsField for fields that support templates, labels, hints, and errors.
+ */
+abstract class InputTagField extends BaseField
+{
+    use InputDataWithCustomNameAndValueTrait;
+    use InputTagMethodsTrait;
+
+    final protected function generateContent(): ?string
+    {
+        return $this->generateInput();
+    }
+
+    abstract protected function generateInput(): string;
+}
